Rename agent protocol to AITP, use query param mockedAitpMessages to include mocked messages
