+++ hub/demo/src/app/api/v1/aitp/request_data/forms/shipping_address_international.json/route.ts
@@ -1,6 +1,6 @@
 import { type z } from 'zod';
 
-import { type requestDataFormSchema } from '~/components/threads/messages/agent-protocol/schema';
+import { type requestDataFormSchema } from '~/components/threads/messages/aitp/schema';
 
 // See "Generic Formats": https://www.uxmatters.com/mt/archives/2008/06/international-address-fields-in-web-forms.php
