+++ hub/demo/src/components/threads/messages/aitp/RequestChoiceProducts.tsx
