+++ hub/demo/src/components/threads/messages/aitp/styles.module.scss
